Temperature Commit #1
I got the temperature to display.  In doing so I had to add in the
Hardware and Sensor classes.  I have been meaning to do this for some
time.  The CounterList and CounterListItem classes will be going away
and the details view will be handled the same way the temperature view
is.